fix: fixed a bug so it doesnt crash if the txt files do no texist

+++ src/utils/saving_messages.py
@@ -1,12 +1,13 @@
-
+import os
 
 def save_message(Status: str, file_name: str):
 
     # clear the file when reached the limit:
-    with open(file_name, 'r', encoding='utf-8') as file:
-        lines = file.readlines()
-        if len(lines) > 500:
-            open(file_name, 'w').close()
+    if os.path.exists(file_name):
+        with open(file_name, 'r', encoding='utf-8') as file:
+            lines = file.readlines()
+            if len(lines) > 500:
+                open(file_name, 'w').close()
 
     #save a message
     with open(file_name, 'a', encoding='utf-8') as file:
